Debounce reload events
For some reason I sometimes see multiple reload events in the example app, for saving a template once. Possibly there’s something in the way watchman watches the project + example app that causes extra events.

This PR fixes that by “debouncing” - starting a 50ms timer when we see a change, and restarting it out every time we get another change.
This is a good idea anyway as it will help workflows where multiple files change in quick succession, e.g. when running Tailwind and changing HTML that triggers a CSS change.